// src/services/coin/coin.types.ts
export interface ICoinPurchaseManagersRequest {
  created_at_start?: string;
  created_at_end?: string;
  search_type?: string;
  search_keyword?: string;
  delng_sttus?: string[];
  page: number;
  per_page: number;
  orderby?: string[];
}